+++ src/main/java/com/team24/newsfeed/exception/ErrorResult.java
@@ -1,7 +1,9 @@
 package com.team24.newsfeed.exception;
 
 import lombok.AllArgsConstructor;
+import lombok.Getter;
 
+@Getter
 @AllArgsConstructor
 public class ErrorResult {
