+++ src/main/java/com/team24/newsfeed/exception/NewsfeedExceptionConst.java
@@ -1,4 +1,4 @@
-package com.team24.newsfeed.exception.profileexception;
+package com.team24.newsfeed.exception;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -6,7 +6,7 @@ import org.springframework.http.HttpStatus;
 
 @Getter
 @AllArgsConstructor
-public enum ProfileExceptionConst {
+public enum NewsfeedExceptionConst {
 
     PASSWORD_MISMATCH(HttpStatus.UNAUTHORIZED, "사용자 비밀번호가 일치하지 않습니다."),
     SAME_PASSWORD(HttpStatus.CONFLICT, "비밀번호는 최소 8자 이상이어야 하며, 영문자, 숫자, 특수문자를 포함해야 합니다."),
