+++ src/main/java/com/team24/newsfeed/exception/profile/NotSamePasswordException.java
@@ -1,0 +1,12 @@
+package com.team24.newsfeed.exception.profile;
+
+import com.team24.newsfeed.exception.NewsfeedException;
+
+import static com.team24.newsfeed.exception.NewsfeedExceptionConst.*;
+
+public class NotSamePasswordException extends NewsfeedException {
+
+    public NotSamePasswordException() {
+        super(PASSWORD_MISMATCH);
+    }
+}
