--- src/main/java/com/team24/newsfeed/exception/profileexception/SameUpdatePasswordException.java
@@ -1,10 +1,0 @@
-package com.team24.newsfeed.exception.profileexception;
-
-import static com.team24.newsfeed.exception.profileexception.ProfileExceptionConst.SAME_PASSWORD;
-
-public class SameUpdatePasswordException extends ProfileException {
-
-    public SameUpdatePasswordException() {
-        super(SAME_PASSWORD);
-    }
-}
