+++ src/main/java/com/team24/newsfeed/service/ProfileService.java
@@ -3,9 +3,9 @@ package com.team24.newsfeed.service;
 import com.team24.newsfeed.domain.User;
 import com.team24.newsfeed.dto.request.ProfileUpdateDto;
 import com.team24.newsfeed.dto.response.ProfileResponseDto;
-import com.team24.newsfeed.exception.profileexception.InvalidPasswordPatternException;
-import com.team24.newsfeed.exception.profileexception.NotSamePasswordException;
-import com.team24.newsfeed.exception.profileexception.SameUpdatePasswordException;
+import com.team24.newsfeed.exception.profile.InvalidPasswordPatternException;
+import com.team24.newsfeed.exception.profile.NotSamePasswordException;
+import com.team24.newsfeed.exception.profile.SameUpdatePasswordException;
 import com.team24.newsfeed.repository.UserRepository;
 import com.team24.newsfeed.security.UserDetailsImpl;
 import org.springframework.security.crypto.password.PasswordEncoder;
